requirements.txt
streamlit==1.26.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.35.0
python-dotenv==1.0.0
numpy==1.26.0